// src/typechain-types/ITokenFactory.ts
export type TokenInfoStruct = {
  exists: boolean;
  id: BigNumberish;
  name: string;
  serviceToken: string;
};